% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configurepython.R
\name{configure_python}
\alias{configure_python}
\title{Sets up a conda environment called castoredc_api for exporting Castor EDC data
If environment already exists, loads the environment and installs necessary packages.}
\usage{
configure_python(libname, pkgname)
}
\description{
Sets up a conda environment called castoredc_api for exporting Castor EDC data
If environment already exists, loads the environment and installs necessary packages.
}
\examples{
\dontrun{
configure_python()
}
}
