% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportdata.R
\name{export_data}
\alias{export_data}
\title{Downloads data from the Castor EDC database}
\usage{
export_data(client_id, client_secret, study_id, url)
}
\arguments{
\item{client_id}{The client id of your Castor account}

\item{client_secret}{The client secret of your Castor account}

\item{study_id}{The id of the study you wish to interact with}

\item{url}{The url where your Castor database resides}
}
\value{
A list containing dataframes with all study, survey and report data.
}
\description{
Downloads data from the Castor EDC database
}
\examples{
\dontrun{
export_data('client_id', 'client_secret', 'study_id', 'data.castoredc.com')
}
}
