% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportdata.R
\name{read_and_delete}
\alias{read_and_delete}
\title{Reads a feather file from path, deletes the file and returns the dataframe.}
\usage{
read_and_delete(path)
}
\arguments{
\item{path}{The path where the feather file resides}
}
\value{
A dataframe read from the feather file
}
\description{
Reads a feather file from path, deletes the file and returns the dataframe.
}
\examples{
\dontrun{
read_and_delete(path_to_file)
}
}
